requirements.txt
absl-py==2.2.2
accelerate @ file:///home/conda/feedstock_root/build_artifacts/accelerate_1741979395147/work
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///D:/bld/aiohttp_1745255715696/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1734342155601/work
astunparse==1.6.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
audioread @ file:///D:/bld/audioread_1725357528890/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///D:/bld/cffi_1725560706917/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///D:/bld/contourpy_1744743081819/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1746740462479/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1706434688412/work
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
flatbuffers==25.2.10
fonttools @ file:///D:/bld/fonttools_1743732204569/work
frozenlist @ file:///D:/bld/frozenlist_1746635344657/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1741403990995/work
gast==0.6.0
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1744136132727/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1746352627188/work
keras==3.9.2
kiwisolver @ file:///D:/bld/kiwisolver_1725459295513/work
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
libclang==18.1.1
librosa @ file:///home/conda/feedstock_root/build_artifacts/librosa_1692209066689/work
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.1
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.1
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
msgpack @ file:///D:/bld/msgpack-python_1725975250448/work
multidict @ file:///D:/bld/multidict_1744972490993/work
multiprocess @ file:///D:/bld/multiprocess_1724954693267/work
munkres==1.1.4
mutagen @ file:///home/conda/feedstock_root/build_artifacts/mutagen_1734757858208/work
namex==0.0.9
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numba @ file:///D:/bld/numba_1744232031776/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp311-cp311-win_amd64.whl#sha256=46eb5850478cda64bda6d2f2d3d6ad8ceeb7cc168b83532e4f9dbbcb42e22ab2
opt_einsum==3.4.0
optree==0.15.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///D:/bld/pandas_1744430491874/work
Pillow @ file:///D:/bld/pillow_1694722928447/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1733421311631/work
propcache @ file:///D:/bld/propcache_1744524926447/work
protobuf==5.29.4
psutil @ file:///D:/bld/psutil_1740663127374/work
pyarrow==16.1.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydub @ file:///home/conda/feedstock_root/build_artifacts/pydub_1734638032476/work
Pygments==2.19.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex @ file:///D:/bld/regex_1730952204165/work
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==14.0.0
safetensors @ file:///D:/bld/safetensors_1740651439698/work
scikit-learn @ file:///D:/bld/scikit-learn_1736496780928/work/dist/scikit_learn-1.6.1-cp311-cp311-win_amd64.whl#sha256=b8f37c282c66c737a09228918d0bfcadafb9f1d4bfb1dc6289040870d453703b
scipy @ file:///C:/bld/scipy-split_1739790799455/work/dist/scipy-1.15.2-cp311-cp311-win_amd64.whl#sha256=7d4ec4c2c958f31576b5ea44433874f567d6f5e65bfcde22e2d960ac86b80515
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
soundfile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1737836266465/work
soxr @ file:///D:/bld/soxr-python_1725345441661/work
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tokenizers @ file:///D:/bld/tokenizers_1713402855834/work/bindings/python
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1714036347648/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///D:/bld/unicodedata2_1736692523011/work
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
Werkzeug==3.1.3
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
wrapt==1.17.2
xxhash @ file:///D:/bld/python-xxhash_1740594813776/work
yarl @ file:///D:/bld/yarl_1744972498902/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
